+++ frontend/src/Components/BestSeller/BestSeller.jsx
@@ -1,45 +1,28 @@
-import React, { useEffect, useState } from "react";
-import Slider from "react-slick";
-import "./bestseller.scss";
-import settings from "../../Hooks/slideSetting";
-import instance from "../../api/api";
+// BestSeller.jsx
+import React from 'react';
+import Slider from 'react-slick';
+import './bestseller.scss';
+import useAPI from '../../Hooks/useAPI'; 
+import settings from '../../Hooks/slideSetting'; 
+
 function BestSeller() {
-  const [bookData, setBookData] = useState([]);
- 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await instance({
-          url: "bestsellers",
-          params: {
-            key: "AIzaSyDmoBeD1zSdaHgR9nh7HUS142-0L6iNL80" 
-          },
-          method: "GET",
-        });
-        setBookData(response.data.items);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      }
-    };
-  
-    fetchData();
-  }, []);
-  
+
+  const { data: bookData, loading, error } = useAPI('bestseller');
+  if (loading) return <p>Loading...</p>;
+  if (error) return <p>Error: {error.message}</p>;
 
   return (
     <div className="allBestSellerContainer">
       <div className="bestSellerContainer">
         <h1>Best Seller</h1>
         <Slider {...settings}>
-          {bookData.map((item) => {
-            return (
-              <div key={item.id} className="bestSeller">
-                {/* <img className='bestSellerPic' src={item.imageLinks.thumbnail} alt="No pic" /> */}
-                <p className="title">{item.volumeInfo.title}</p>
-                <p className="author">{item.volumeInfo.authors}</p>
-              </div>
-            );
-          })}
+          {bookData?.items.map((item) => (
+            <div key={item.id} className="bestSeller">
+              <img className="bestSellerPic" src={item.volumeInfo.imageLinks?.thumbnail} alt="No pic" />
+              <p className="title">{item.volumeInfo.title}</p>
+              <p className="author">{item.volumeInfo.authors?.join(', ')}</p>
+            </div>
+          ))}
         </Slider>
       </div>
       <a className="seeAll" href="#">
