+++ frontend/src/Components/InternationalBestSeller/InternationalbestSeller.jsx
@@ -1,31 +1,13 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import Slider from "react-slick";
 import settings from "../../Hooks/slideSetting";
 import '../BestSeller/bestseller.scss'
-import instance from '../../api/api';
+import useAPI from '../../Hooks/useAPI';
 
 function InternationalBestSeller() {
-  const [bookData,setBookData] = useState([])
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await instance({
-          url: "internationalbestseller",
-          params: {
-            key: "AIzaSyDmoBeD1zSdaHgR9nh7HUS142-0L6iNL80" 
-          },
-          method: "GET",
-        });
-    
-        setBookData(response.data.items);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      }
-    };
-  
-    fetchData();
-  }, []);
-  
+  const { data: bookData, loading, error } = useAPI("international");
+  if (loading) return <p>Loading...</p>;
+  if (error) return <p>Error: {error.message}</p>;
 
   return (
     <div className='allBestSellerContainer'>
@@ -33,7 +15,7 @@ function InternationalBestSeller() {
             <h1>International BestSeller</h1>
             <Slider {...settings}>
                 {
-                  bookData.map((item)=>{
+                  bookData?.items.map((item)=>{
                     return (
                       <div key={item.id} className='bestSeller'>
                           <img className='bestSellerPic' src={item.volumeInfo.imageLinks.thumbnail} alt="No pic" />
