+++ frontend/src/Components/Navbar/navbar.scss
@@ -18,7 +18,7 @@ $menu-hover-color: #555;
     justify-content: space-between;
     padding: $navbar-padding;
     position: fixed;
-    z-index: 1;
+    z-index: 1000;
     top: 0;
     left: 0;
     background-color:rgba(255, 255, 255, 0.918);
