+++ frontend/src/Components/NewArrival/NewArrival.jsx
@@ -1,38 +1,22 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import "../BestSeller/bestseller.scss";
 import settings from "../../Hooks/slideSetting";
-import instance from "../../api/api";
+import useAPI from "../../Hooks/useAPI";
 function BestSeller() {
-  const [bookData, setBookData] = useState([]);
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await instance({
-          url: "newarrival",
-          params: {
-            key: "AIzaSyDmoBeD1zSdaHgR9nh7HUS142-0L6iNL80" 
-          },
-          method: "GET",
-        });
-       
-        setBookData(response.data.items);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      }
-    };
-  
-    fetchData();
-  }, []);
-  
+
+  const { data: bookData, loading, error } = useAPI("international");
+  if (loading) return <p>Loading...</p>;
+  if (error) return <p>Error: {error.message}</p>;
+
   return (
     <div className="allBestSellerContainer">
       <div className="bestSellerContainer">
         <h1>New Arrival</h1>
         <Slider {...settings}>
-          {bookData.map((item) => {
+          {bookData?.items.map((item) => {
             return (
               <div key={item.id} className="bestSeller">
                 <img
