+++ frontend/src/Hooks/useAPI.jsx
@@ -1,0 +1,32 @@
+import { useState, useEffect } from 'react';
+import instance from '../api/api'; 
+
+const useAPI = (url, options = {}) => {
+  const [data, setData] = useState(null);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await instance({
+          url,
+          ...options,
+          method: options.method || 'GET',
+        });
+        setData(response.data);
+        console.log(response.data)
+      } catch (error) {
+        setError(error);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchData();
+  }, [url]);
+
+  return { data, loading, error };
+};
+
+export default useAPI;
