+++ frontend/src/Hooks/useAuthentication.jsx
@@ -1,14 +1,29 @@
+// useAuthentication.jsx
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import {
   signInWithPopup,
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+  signOut as firebaseSignOut,
 } from "firebase/auth";
-import { auth, provider } from "../lib/firebase";
+import { auth, provider, db } from "../lib/firebase";
 
 const useAuthentication = () => {
   const [user, setUser] = useState(null);
   const [error, setError] = useState(null);
   const navigate = useNavigate();
+
+  useEffect(() => {
+    const unsubscribe = auth.onAuthStateChanged((user) => {
+      if (user) {
+        setUser(user);
+      } else {
+        setUser(null);
+      }
+    });
+    return () => unsubscribe();
+  }, []);
 
   const signInWithGoogle = async () => {
     try {
@@ -22,28 +37,46 @@ const useAuthentication = () => {
     }
   };
 
+  const signInWithEmailAndPasswordFunc = async (email, password) => {
+    try {
+      const userCredential = await signInWithEmailAndPassword(auth, email, password);
+      setUser(userCredential.user);
+      return true;
+    } catch (err) {
+      setError(err.message);
+      return false;
+    }
+  };
+
+  const createUserWithEmailAndPasswordFunc = async (email, password) => {
+    try {
+      const userCredential = await createUserWithEmailAndPassword(auth, email, password);
+      const user = userCredential.user;
+      await db.collection("users").doc(user.uid).set({
+        email: user.email,
+        uid: user.uid,
+      });
+      setUser(user);
+      return true;
+    } catch (err) {
+      setError(err.message);
+      return false;
+    }
+  };
+
   const signOut = async () => {
     setUser(null);
     await localStorage.removeItem("gmail");
-    auth.signOut();
+    await firebaseSignOut(auth);
     navigate("/login");
   };
-
-  useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged((user) => {
-      if (user) {
-        setUser(user);
-      } else {
-        setUser(null);
-      }
-    });
-    return () => unsubscribe();
-  }, []);
 
   return {
     user,
     error,
     signInWithGoogle,
+    signInWithEmailAndPassword: signInWithEmailAndPasswordFunc,
+    createUserWithEmailAndPassword: createUserWithEmailAndPasswordFunc,
     signOut,
   };
 };
